apps/data_ingestor/pyproject.toml
[project]
name = "data-ingestor"
version = "0.1.0"
description = "An ingestion service for collecting data from external sources."
authors = [{ name = "Ben Ledingham" }]
dependencies = [
    "requests",
    "pytest-cov",
    "python-dotenv",
    "ruff",
    "pymongo",
]
requires-python = ">=3.10"